+++ backend/dataall/modules/datasets/api/__init__.py
@@ -1,7 +1,7 @@
 """The GraphQL schema of datasets and related functionality"""
 from dataall.modules.datasets.api import (
     table_column,
-    DatasetProfiling
+    profiling
 )
 
-__all__ = ["table_column", "DatasetProfiling"]
+__all__ = ["table_column", "profiling"]
